backend/requirements.txt
# Backend Dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
pandas==2.2.2
anthropic>=0.40.0
httpx>=0.27.0

# Database
# sqlite3 is built-in to Python
